fix: remove launch helper dependencies on CLI and grid execution
