+++ src/hyperloom/inference_optimizer/tests/test_grid_runner_helpers_coverage_unit.py
@@ -13,6 +13,7 @@ import pytest
 import yaml
 
 from hyperloom.orchestrator.actions.executors import _grid_runner as gr
+from hyperloom.orchestrator.actions.executors import _benchmark_interpreter as bi
 
 # Patch compatibility-filter helpers in the ``_grid_variant_filter`` sibling, where apply_compatibility_filter
 # resolves them (not via the re-export).
@@ -156,7 +157,7 @@ def test_probe_swallows_subprocess_error(monkeypatch) -> None:
 # -- _resolve_probe_python / probe interpreter selection ------------------
 def test_resolve_probe_python_prefers_magpie_interpreter(monkeypatch) -> None:
     # The harness interpreter is used directly; no vllm-exe resolution is tried.
-    monkeypatch.setattr(gr, "_resolve_magpie_python", lambda: "/srv/venv/bin/python")
+    monkeypatch.setattr(bi, "_resolve_magpie_python", lambda: "/srv/venv/bin/python")
     monkeypatch.setattr(
         gr.shutil, "which", lambda *_a, **_k: (_ for _ in ()).throw(AssertionError("which() should not be called"))
     )
@@ -165,7 +166,7 @@ def test_resolve_probe_python_prefers_magpie_interpreter(monkeypatch) -> None:
 
 def test_resolve_probe_python_falls_back_to_vllm_venv(monkeypatch) -> None:
     # magpie_python is the canonical default -> pin the venv that backs ``vllm serve``.
-    monkeypatch.setattr(gr, "_resolve_magpie_python", lambda: "/opt/venv/bin/python")
+    monkeypatch.setattr(bi, "_resolve_magpie_python", lambda: "/opt/venv/bin/python")
     monkeypatch.setattr(gr.shutil, "which", lambda name: "/other/venv/bin/vllm" if name == "vllm" else None)
     monkeypatch.setattr(gr.os.path, "exists", lambda p: p == "/other/venv/bin/python")
     assert gr._resolve_probe_python() == "/other/venv/bin/python"
@@ -173,7 +174,7 @@ def test_resolve_probe_python_falls_back_to_vllm_venv(monkeypatch) -> None:
 
 def test_resolve_probe_python_no_bare_python3_fallback(monkeypatch) -> None:
     # With no resolvable vllm exe, fall back to the canonical magpie default — never a bare "python3".
-    monkeypatch.setattr(gr, "_resolve_magpie_python", lambda: "/opt/venv/bin/python")
+    monkeypatch.setattr(bi, "_resolve_magpie_python", lambda: "/opt/venv/bin/python")
     monkeypatch.setattr(gr.shutil, "which", lambda *_a, **_k: None)
     assert gr._resolve_probe_python() == "/opt/venv/bin/python"
