+++ src/hyperloom/orchestrator/actions/executors/baseline.py
@@ -4205,7 +4205,11 @@ __all__ = [
     "BASELINE_COLD_START_TIMEOUT_SEC",
     "AGENTX_BASELINE_OVERHEAD_SEC",
     "AGENTX_DEFAULT_DURATION_SEC",
+    "AGENTX_CANON_WARMUP_GRACE_SEC",
+    "AGENTX_CANON_WARMUP_CONC",
     "BASELINE_DEFAULT_TIMEOUT_SEC",
+    "agentx_warmup_grace_conc",
+    "agentx_warmup_grace_sec",
     "agentx_baseline_timeout_sec",
     "BaselineExecutor",
     "COLD_START_KERNEL_THRESHOLD",
